requirements.txt
kivy
Pillow
docutils
pygments
pypiwin32
kivy.deps.sdl2
kivy.deps.glew
kivy.deps.gstreamer
kivy.deps.angle
